refactor: Split pruning BFS rewrite into .cpp and .hpp files
